New templates to generate a pdf ( Issue#7 ) file through latex code.
As well as for the html generator, we need new templates for the
latex generator, and then this code will be compiled to generate
the pdf file.
In assets, there are two files, one for the packages needed to compile
the latex code, and the other includes definition of new commands
in order to make the latex code more legible and usable.