Table environment has been removed from latex code generated for tables.
The elements generated for the pdf file ( Issue #7 ) are included in
minipage environments. For that, tables that before was included in
table environments has been change to generate the tabular code
without the environment. It is also a better way to express the
code of the table elements, due to environment is out of scope (is
more related to the end user to decide which environment he wants
to use).