Some minor bugs corrections.
The name testWilcoxon was still used instead of testPaired. A file
has been renamed and also the calls for respective functions and
the functions themeselves.

Also the historic was not properly generated in the case of expExtract.
In that case, a variable was not properly used (now the ued
information is in tags).